SPEED-978 Updating the Go Test Bench Environment to use Official Contrast Service Container
This change adds the officially supported contrast-service container to the demo compose file
for use by the Go Agent as well as adds the necessary volume configuration so that the container
can properly load the configuration file without the need to build the config into the image
directly since building an image with a config file internally doesn't make sense.

This change also corrects an improper ListenAndServe entry where `localhost` was passed as the
hostname which was breaking the docker runtime network mapping for the ports in the compose file.

This change also updates the agent dockerfile to remove the build dependency on the contrast_security.yaml
moving it to a mounted volume similar the method used by the contrast-service image.