GO-1080: Add scripts to fully exercise the app
TODO: This is a temporary commit for the team to be able to use
the script to test other changes. I will refactor the code so that
common endpoints reuse the same request generators before making a
PR. For now, you can test the changes with:
  go run ./cmd/std/exercise -addr localhost:8080
There's an analogous command in ./cmd/gin/exercise.