GO-1299: Adding go-swagger to the test bench with cmd injection
These changes add a go-swagger application to the test bench.
This is done by adding a new module inside the /cmd directory
which is intended to make a new executable which can create
and run a swagger server containing sample vulnerabilities.

With the current changes, the first vulnerability we are adding
is command injection through user supplied URL query.

To exercise and view the cmd injection vulnerability:
1) launch the swagger app with contrast-go from the root of the repo:
	contrast-go run ./cmd/go-swagger
2) provide input in the Query section of the cmd injection page.

Routes, paths and the front-end have the overall structure of gin and
std, new vulnerabilities can be added to the server by adding a handler
of the type provided by swagger's middleware package to handle the given
endpoint containing the vulnerability and then using the api.HTMLProducer
in order to handle UI construction.