GO-1036: Remove unused NoSQLi code
The NoSQLi vulnerability endpoint has been unused for well over
a year now. Two new frameworks (gin and go-swagger) have been added
without NoSQLi endpoints. "NoSQLi" vulnerabilities are also a bit
of a misnomer, since there isn't a NoSQL language. The injection
vulnerabilities are tied to specific databases, and often specific
libraries. We should really move toward MongoDB NoSQLi endpoints,
or DynamoDB NoSQLi endpoints, etc. Lastly, the existing code to
trigger a MongoDB NoSQLi vulnerability requires a MongoDB
instance to be running alongside the app. No other
vulnerabilities depend on external services, so using this code
will increase the operational overhead of running the app.

All that said, the current code is unused and we haven't heard
feedback that it should be added. To avoid unnecessary work, this
change removes the NoSQLi code entirely so that we can have a
clean slate when we add NoSQLi support in the future.