GO-1959: change default cmdi payload to a valid command (#57)
Originally, the default cmdi payload did not form a valid command
when using the unsafe handler. This was fine for assess, but not for
protect. Protect will only block if the command tries to run. This
would not happen if the command name (previously "hello") was not
found in the PATH. To avoid confusing end users, the default cmdi
payload has been changed to a valid command.