pipelines: fix broken pipelines from previous commit (#62)
The last PR had a merge conflict with main. I thought I had it
resolved, but it turns out I missed one. I'm honestly surprised that
both VSCode and Github let me merge a PR that still had a conflict. I
guess the text highlighting the conflict was just naively added to
the workflow file. This broke the file of course, but it still let me
merge. Very strange. Anyway, let's fix this.