GO-2278: streamline docker setup and update documentation
The docker setup here is not complex enough to warrant docker-compose
files, so this commit removes them in favor of simple `docker build`
and `docker run` commands.

The other change here is to the docker image version specified. It is
too easy to forget to change the hard-coded version when we update
our support window. We have changed our new version support process a
lot since we last updated the version here. We now tend to support
new Go releases very soon after they release, so I think we can get
away with having the default version be `latest` in our Dockerfiles.

This commit updates the documentation to match both the above changes
and changes made to teamserver.